feat(create listing): add create listing functionality
added onCreatePost that handles form submission and input data

added createPost that uses POST to update data to API

added createListingTemplate that dynamically generates a listing template for creating a post on click of 'create post' button.

added and init function that renders the template and adds a form submit listener.

#14 #12